requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
Cython
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
gast==0.3.3
google-auth==1.10.1
google-auth-oauthlib==0.4.1
grpcio==1.26.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.4.0
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterthemes==0.20.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lesscpy==0.13.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
mock==3.0.5
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.3
notebook==6.0.2
numpy
oauthlib==3.1.0
pandas
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
tb-nightly==2.2.0a20200106
tensorboard
tensorflow-estimator
# tensorflow-gpu==1.14.0
tensorflow==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
zipp==1.0.0
logbook
raven
psutil
implicit
